uhd: fix seg fault at startup
On the E100, and possibly other systems, OpenBTS seg faults
at startup. The cause is a null reference to the global
logging file descriptor occuring due to an initializer order
dependency, which is not guaranteed across compilation
units.

When this occurs, check for for null and set the output to
stdout.

Signed-off-by: Thomas Tsou <ttsou@vt.edu>