usrp: remove unnecessary I/Q swap for non-uhd
This bug slipped in because the following patch
was made before tuning adjustment changes that
negated the need for the I/Q swap.

753118031e633cd4957d15369169f5b6fe8aec39

At this point, we assume that there are no
big-endian users of OpenBTS with the USRP1. If
there are any SPARC, PowerPC, or MIPS users,
hopefully they will be resourceful enough to
search the commit log and add the necessary
byte swaps.

Signed-off-by: Thomas Tsou <ttsou@vt.edu>